+++ code/ext/ext.lib.automerge/src/crdt.cmd/Cmd.impl.ts
@@ -1,4 +1,4 @@
-import { listenerFactory } from './Cmd.Listener';
+import { listenerFactory } from './u.Listener';
 import { DEFAULTS, ObjectPath, Time, type t } from './common';
 import { Events, Is, Path } from './u';
